move copy to clipboard action to common module
