+++ common-lib/src/main/java/com/google/cloud/tools/intellij/ui/CopyToClipboardActionListener.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2017 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.cloud.tools.intellij.ui;
+
+import java.awt.Toolkit;
+import java.awt.datatransfer.Clipboard;
+import java.awt.datatransfer.StringSelection;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+/** {@link ActionListener} that copies text to the clipboard on action performed. */
+public class CopyToClipboardActionListener implements ActionListener {
+  private final String text;
+
+  public CopyToClipboardActionListener(String text) {
+    this.text = text;
+  }
+
+  @Override
+  public void actionPerformed(ActionEvent event) {
+    Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
+    clipboard.setContents(new StringSelection(text), null /*owner*/);
+  }
+}
